Update #13 - Big FPS boost for most players
The hotbar, chat, and debug overlays in the game have been optimized and now only require a fraction as much CPU/GPU usage to be displayed on screen

+++ sources/main/java/net/lax1dude/eaglercraft/v1_8/opengl/FixedFunctionShader.java
@@ -22,7 +22,7 @@ class FixedFunctionShader {
 		static final String FILENAME_FSH = "/assets/eagler/glsl/core.fsh";
 
 		static final String PRECISION_INT = "lowp";
-		static final String PRECISION_FLOAT = "highp";
+		static final String PRECISION_FLOAT = "mediump";
 		static final String PRECISION_SAMPLER = "lowp";
 		
 		static final String MACRO_ATTRIB_TEXTURE = "COMPILE_TEXTURE_ATTRIB";
